6336 | 57 files | M DumpSource2/schemas/client/CDOTA_Ability_Grimstroke_SoulChain.h,  M DumpSource2/schemas/client/CDOTA_Ability_Snapfire_MortimerKisses.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Batrider_Firefly.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DeathProphet_Scepter.h,  M DumpSource2/schemas/client/CDOTA_Modifier_EarthSpirit_Magnetize.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Elder_Titan_AncestralSpirit.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Elder_Titan_EarthSplitter_Caster.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Furbolg_Enrage_Damage_OnDeath.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Grimstroke_SoulChain.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Hoodwink_Sharpshooter_VisionThinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Invoker_IceWall_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Kunkka_Tidebringer.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Kunkka_Torrent_Storm.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Mars_Bulwark_Active.h,  M DumpSource2/sch
https://steamdb.info/patchnotes/17519299/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Batrider_Firefly.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_Batrider_Firefly
 {
 	bool m_bActive;
-	int32 radius;
+	float32 radius;
 	int32 m_ifirefly_Active;
-	int32 tree_radius;
-	int32 damage_per_second;
-	int32 movement_speed;
+	float32 tree_radius;
+	float32 damage_per_second;
+	float32 movement_speed;
 	int32 bonus_slow_resistance;
 	float32 tick_interval;
 	ParticleIndex_t m_nFXIndex;
