+++ DumpSource2/schemas/client/CDOTA_Modifier_Elder_Titan_AncestralSpirit.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Elder_Titan_AncestralSpirit
 {
-	int32 pass_damage;
-	int32 radius;
+	float32 pass_damage;
+	float32 radius;
 	int32 controllable;
 	int32 m_nCreepsHit;
 	int32 m_nHeroesHit;
