+++ DumpSource2/schemas/client/CDOTA_Modifier_Kunkka_Tidebringer.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_Kunkka_Tidebringer
 {
 	bool m_bTidebringerAttack;
-	int32 cleave_starting_width;
-	int32 cleave_ending_width;
-	int32 cleave_distance;
-	int32 damage_bonus;
+	float32 cleave_starting_width;
+	float32 cleave_ending_width;
+	float32 cleave_distance;
+	float32 damage_bonus;
 	ParticleIndex_t m_nFXIndex;
-	int32 cleave_damage;
+	float32 cleave_damage;
 	float32 cooldown_reduction_per_hero_hit;
 };
