+++ DumpSource2/schemas/client/CDOTA_Modifier_Kunkka_Torrent_Storm.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Kunkka_Torrent_Storm
 {
-	int32 torrent_min_distance;
-	int32 torrent_max_distance;
+	float32 torrent_min_distance;
+	float32 torrent_max_distance;
 	int32 m_iExplosionQuadrant;
 };
