+++ DumpSource2/schemas/client/CDOTA_Modifier_Mars_Bulwark_Active.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_Mars_Bulwark_Active
 {
 	int32 redirect_chance;
-	int32 redirect_range;
+	float32 redirect_range;
 	float32 forward_angle;
 	float32 side_angle;
-	int32 redirect_speed_penatly;
-	int32 redirect_close_range;
+	float32 redirect_speed_penatly;
+	float32 redirect_close_range;
 	int32 attack_redirection_grace_distance;
 	int32 m_nPoseParameterWE;
 	int32 m_nPoseParameterNS;
