+++ DumpSource2/schemas/client/CDOTA_Modifier_Pudge_GraftFlesh.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Pudge_GraftFlesh
 {
-	int32 flesh_heap_range;
+	float32 flesh_heap_range;
 	float32 flesh_heap_strength_buff_amount;
 };
