+++ DumpSource2/schemas/client/CDOTA_Modifier_Shredder_ReactiveArmor_Bomb.h
@@ -6,8 +6,8 @@ class CDOTA_Modifier_Shredder_ReactiveArmor_Bomb
 	float32 shield_per_sec_per_enemy;
 	float32 duration;
 	float32 base_explosion;
-	int32 radius;
-	int32 explosion_radius;
+	float32 radius;
+	float32 explosion_radius;
 	int32 m_nDamageAbsorbed;
 	GameTime_t m_timeLastTick;
 	GameTime_t m_StartTime;
