+++ DumpSource2/schemas/client/CDOTA_Modifier_Tinker_MarchOfTheMachinesThinker.h
@@ -1,9 +1,9 @@
 class CDOTA_Modifier_Tinker_MarchOfTheMachinesThinker
 {
-	int32 collision_radius;
-	int32 radius;
-	int32 distance;
-	int32 speed;
+	float32 collision_radius;
+	float32 radius;
+	float32 distance;
+	float32 speed;
 	int32 machines_per_sec;
 	int32 heal_per_second;
 	Vector m_vDir;
