+++ DumpSource2/schemas/client/CDOTA_Modifier_Viper_CorrosiveSkin_Slow.h
@@ -1,9 +1,9 @@
 class CDOTA_Modifier_Viper_CorrosiveSkin_Slow
 {
 	int32 bonus_attack_speed;
-	int32 damage;
-	int32 effect_multiplier_distance;
-	int32 effect_multiplier;
+	float32 damage;
+	float32 effect_multiplier_distance;
+	float32 effect_multiplier;
 	int32 bonus_status_resistance;
 	int32 nethertoxin_bonus_pct;
 	bool m_bEffectMultiplied;
