+++ DumpSource2/schemas/client/CDOTA_Modifier_Zuus_Cloud.h
@@ -2,5 +2,5 @@ class CDOTA_Modifier_Zuus_Cloud
 {
 	CountdownTimer m_CloudBoltTimer;
 	float32 cloud_bolt_interval;
-	int32 cloud_radius;
+	float32 cloud_radius;
 };
