+++ DumpSource2/schemas/client/C_DOTA_Ability_EarthSpirit_Magnetize.h
@@ -1,6 +1,6 @@
 class C_DOTA_Ability_EarthSpirit_Magnetize
 {
-	int32 cast_radius;
-	int32 rock_explosion_radius;
+	float32 cast_radius;
+	float32 rock_explosion_radius;
 	float32 damage_duration;
 };
