+++ DumpSource2/schemas/server/CDOTA_Ability_Grimstroke_SoulChain.h
@@ -1,6 +1,6 @@
 class CDOTA_Ability_Grimstroke_SoulChain
 {
 	float32 chain_duration;
-	int32 chain_latch_radius;
+	float32 chain_latch_radius;
 	float32 creep_duration_pct;
 };
