+++ DumpSource2/schemas/server/CDOTA_Ability_Invoker_Tornado.h
@@ -1,11 +1,11 @@
 class CDOTA_Ability_Invoker_Tornado
 {
-	int32 vision_distance;
+	float32 vision_distance;
 	float32 end_vision_duration;
 	float32 lift_duration;
 	float32 base_damage;
 	float32 quas_damage;
 	float32 wex_damage;
 	float32 twister_duration;
-	int32 twister_distance_interval;
+	float32 twister_distance_interval;
 };
