+++ DumpSource2/schemas/server/CDOTA_Ability_QueenOfPain_SonicWave.h
@@ -3,6 +3,6 @@ class CDOTA_Ability_QueenOfPain_SonicWave
 	GameTime_t m_fStartTime;
 	float32 m_fTotalTime;
 	ParticleIndex_t m_nFXIndex;
-	int32 starting_aoe;
-	int32 final_aoe;
+	float32 starting_aoe;
+	float32 final_aoe;
 };
