+++ DumpSource2/schemas/server/CDOTA_Ability_TrollWarlord_WhirlingAxes_Ranged.h
@@ -2,10 +2,10 @@ class CDOTA_Ability_TrollWarlord_WhirlingAxes_Ranged
 {
 	Vector m_vStartPos;
 	int32 m_iArrowProjectile;
-	int32 axe_width;
+	float32 axe_width;
 	float32 axe_speed;
 	float32 axe_range;
-	int32 axe_spread;
+	float32 axe_spread;
 	int32 axe_count;
 	CUtlVector< CHandle< CBaseEntity > > m_hHitUnits;
 };
