+++ DumpSource2/schemas/server/CDOTA_Modifier_DeathProphet_Scepter.h
@@ -1,11 +1,11 @@
 class CDOTA_Modifier_DeathProphet_Scepter
 {
-	int32 radius;
-	int32 spirit_speed;
-	int32 max_distance;
-	int32 give_up_distance;
-	int32 min_damage;
-	int32 max_damage;
+	float32 radius;
+	float32 spirit_speed;
+	float32 max_distance;
+	float32 give_up_distance;
+	float32 min_damage;
+	float32 max_damage;
 	int32 heal_percent;
 	int32 spirit_duration;
 	GameTime_t m_fStartTime;
