+++ DumpSource2/schemas/server/CDOTA_Modifier_Hoodwink_Sharpshooter_VisionThinker.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Hoodwink_Sharpshooter_VisionThinker
 {
-	int32 arrow_width;
+	float32 arrow_width;
 	bool vision_along_indicator;
-	int32 vision_radius_buffer;
+	float32 vision_radius_buffer;
 };
